+++ CHANGELOG.md
@@ -10,4 +10,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 - Adding printable method for mult-armed-bandit class
 - Adding constructor for mult-armed-bandit class
 - Adding pull\_arm function for mult-armed-bandit object
-
+- Adding function that always choose the best arm to pull